Split repository.ts: extract suggestions + budgets/privacy (11/N)
Eleventh slice of the repository.ts carve-up (audit L14):

- repository/suggestions.ts (171) — the member-submitted bot-improvement queue
  (#46) admins triage via list_suggestions / resolve_suggestion.
- repository/budgetsPrivacy.ts (363) — reply budgets, retention purges and the
  member-facing privacy surface (my_data, forget_me, purge_user_data), including
  the deletion-coherence rules that make a purge actually erase a member's
  footprint, digests included.

Also moves the private `invalidateDigestsForInteractions` helper into
repository/shared.ts (exported). It bridges interactions and digests, is called
from four places across what are now different modules, and duplicating it would
risk the two copies diverging on deletion coherence — a privacy-relevant rule.
Verified the body is byte-identical after accounting for the added `export`
keyword; only its location and visibility changed.

budgetsPrivacy imports resolveLinkedIdentities (members.ts) and
getResponseStyle/ResponseStyle (preferences.ts) — purging spans a person's
linked identities and my_data reports their standing style. Both are one-way
domain imports, no cycles.

Also tidies the stale comment #819's review flagged: it still claimed "the
knowledge queries below use this floor" after the knowledge domain moved out and
that import was narrowed.

15/16 moved symbols byte-identical (the 16th is the helper above, identical
modulo `export`). repository.ts is now 2,593 lines, down from 7,137 — a 64%
reduction across eleven slices.
