+++ src/storage/repository.ts
@@ -32,10 +32,10 @@ export * from './repository/preferences.js';
 export * from './repository/memberNotes.js';
 export * from './repository/shared.js';
 export * from './repository/devTeamWatches.js';
-// `export *` does not bind names into this module's own scope, and the
-// knowledge queries below use this floor directly — so import it explicitly.
-import { type Queryable } from './repository/shared.js';
-import { AUTO_ENROLL_ACTOR, resolveLinkedIdentities } from './repository/members.js';
+// `export *` does not bind names into this module's own scope, so anything
+// still living here that uses an extracted symbol must import it explicitly.
+import { invalidateDigestsForInteractions } from './repository/shared.js';
+import { AUTO_ENROLL_ACTOR } from './repository/members.js';
 import { sumShortcutHits } from './repository/shortcutHits.js';
 export * from './repository/accessRequests.js';
 export * from './repository/contextDigests.js';
@@ -51,11 +51,8 @@ export * from './repository/memberProjects.js';
 export * from './repository/members.js';
 export * from './repository/knowledge.js';
 export * from './repository/knowledgeCandidates.js';
-
-// `export *` re-exports for CALLERS but does not bind the names in this
-// module's own scope, so anything still living here that uses an extracted
-// function (getMyData reads the caller's standing style) imports it explicitly.
-import { getResponseStyle, type ResponseStyle } from './repository/preferences.js';
+export * from './repository/suggestions.js';
+export * from './repository/budgetsPrivacy.js';
 
 export interface InteractionInput {
   platform: Platform;
@@ -341,34 +338,6 @@ export async function recentConversationTail(
 }
 
 /**
- * Invalidate every context digest whose provenance refs include any of
- * `interactionIds`, deleting its still-*pending* knowledge candidates first
- * (the same deletion-coherence logic `purgeSingleIdentity` applies, issues
- * #51/#102). Shared so the delete/edit-honouring path (issue #48) invalidates
- * digests built over a message the same way a privacy purge does — otherwise a
- * deleted/edited message's content lives on inside a digest summary. Returns
- * the number of pending candidates removed. No-op on an empty id list.
- */
-async function invalidateDigestsForInteractions(
-  interactionIds: number[],
-  db: Queryable = pool,
-): Promise<number> {
-  if (interactionIds.length === 0) return 0;
-  const { rows: invalidatedDigests } = await db.query(
-    `SELECT id FROM context_digests WHERE example_refs && $1::bigint[]`,
-    [interactionIds],
-  );
-  const digestIds = invalidatedDigests.map((r) => Number(r.id));
-  if (digestIds.length === 0) return 0;
-  const { rowCount: deletedCandidates } = await db.query(
-    `DELETE FROM knowledge_candidates WHERE digest_id = ANY($1::bigint[]) AND status = 'pending'`,
-    [digestIds],
-  );
-  await db.query(`DELETE FROM context_digests WHERE id = ANY($1::bigint[])`, [digestIds]);
-  return deletedCandidates ?? 0;
-}
-
-/**
  * Honour a platform-level message deletion (issue #48): hard-delete the
  * stored copy. Scoped to `(platform, conversationId, messageId)` — message
  * ids are only unique *within* a conversation on some platforms (WhatsApp
@@ -643,369 +612,6 @@ export async function isKnownMessage(
     [platform, conversationId, messageId],
   );
   return rows.length > 0;
-}
-
-// --- Budgets / privacy --------------------------------------------------------
-
-/**
- * Agent replies sent to this user in the last `sinceHours` hours, aggregated
- * across every identity linked to them via `link_member` (so the daily reply
- * budget can't be double-dipped by messaging from a linked Discord account
- * and WhatsApp number instead of one).
- */
-export async function countRepliesToUser(
-  platform: Platform,
-  userId: string,
-  sinceHours = 24,
-): Promise<number> {
-  const identities = await resolveLinkedIdentities(platform, userId);
-  const params: unknown[] = [String(sinceHours)];
-  const conditions = identities.map((id) => {
-    params.push(id.platform, id.userId);
-    return `(platform = $${params.length - 1} AND meta->>'replyToUserId' = $${params.length})`;
-  });
-  const { rows } = await pool.query(
-    `SELECT count(*) AS n FROM interactions
-      WHERE direction = 'outbound'
-        AND created_at > now() - ($1 || ' hours')::interval
-        AND (${conditions.join(' OR ')})`,
-    params,
-  );
-  return Number(rows[0]?.n ?? 0);
-}
-
-/**
- * Delete one identity's stored data — the single-identity core of
- * `purgeUserData`. Runs every delete inside ONE transaction (issue: a crash
- * partway used to leave, e.g., digests alive over already-deleted interactions
- * that a retry could never re-find), mirroring the sibling `linkMembers`/
- * `unlinkMember` pattern.
- */
-async function purgeSingleIdentity(platform: Platform, userId: string): Promise<number> {
-  const client = await pool.connect();
-  try {
-    await client.query('BEGIN');
-
-    // Clear session continuity FIRST, while the user's interactions still
-    // exist for the subquery to find the conversations they were active in.
-    // Without this the `sessions` row keeps mapping the conversation to a live
-    // Claude transcript that still contains the purged messages, so another
-    // member could ask the bot to recall them for up to
-    // SESSION_MAX_TURNS/AGE_HOURS. Same primitive as `clearUserSessions`, run
-    // in-transaction and before the interactions delete below.
-    await client.query(
-      `UPDATE sessions
-          SET claude_session_id = NULL, updated_at = now()
-        WHERE platform = $1
-          AND claude_session_id IS NOT NULL
-          AND conversation_id IN (
-            SELECT DISTINCT conversation_id FROM interactions
-             WHERE platform = $1 AND user_id = $2
-          )`,
-      [platform, userId],
-    );
-
-    const { rows: deletedInteractions } = await client.query(
-      `DELETE FROM interactions
-        WHERE platform = $1
-          AND (user_id = $2 OR (direction = 'outbound' AND meta->>'replyToUserId' = $2))
-        RETURNING id`,
-      [platform, userId],
-    );
-    const messages = deletedInteractions.length;
-    // Deletion coherence (issues #51/#102): a context digest whose summary was
-    // built over any purged interaction is invalidated outright — the next
-    // builder run regenerates the topic without this person's signal. Shared
-    // with the delete/edit-honouring path via `invalidateDigestsForInteractions`.
-    const candidates = await invalidateDigestsForInteractions(
-      deletedInteractions.map((r) => Number(r.id)),
-      client,
-    );
-    // knowledge has no platform column, so this keys on source_user_id alone.
-    // Safe because Discord snowflakes (17-20 digits) and WhatsApp E.164 numbers
-    // (7-15 digits) can't collide as strings (enforced by normalizeMemberId), so
-    // this never touches another platform's user. If that validation loosens, add
-    // a platform column to knowledge and filter on it here.
-    const { rowCount: knowledge } = await client.query(`DELETE FROM knowledge WHERE source_user_id = $1`, [
-      userId,
-    ]);
-    const { rowCount: reports } = await client.query(
-      `DELETE FROM content_reports WHERE platform = $1 AND reporter_user_id = $2`,
-      [platform, userId],
-    );
-    const { rowCount: roster } = await client.query(
-      `DELETE FROM server_roster WHERE platform = $1 AND user_id = $2`,
-      [platform, userId],
-    );
-    const { rowCount: notes } = await client.query(
-      `DELETE FROM member_notes WHERE platform = $1 AND user_id = $2`,
-      [platform, userId],
-    );
-    const { rowCount: suggestions } = await client.query(
-      `DELETE FROM suggestions WHERE platform = $1 AND user_id = $2`,
-      [platform, userId],
-    );
-    // admin_digest_sends (issue #97) is keyed on the same (platform, user id)
-    // identity — purge coherence for an offboarded admin.
-    const { rowCount: digestSends } = await client.query(
-      `DELETE FROM admin_digest_sends WHERE platform = $1 AND platform_user_id = $2`,
-      [platform, userId],
-    );
-    // response_style_prefs (issue #126) is keyed the same way — purge coherence
-    // for anyone who opted into the plain-language preference.
-    const { rowCount: responseStyle } = await client.query(
-      `DELETE FROM response_style_prefs WHERE platform = $1 AND user_id = $2`,
-      [platform, userId],
-    );
-    // language_prefs (issue #189) is keyed the same way — purge coherence for
-    // anyone who opted into a standing language preference.
-    const { rowCount: languagePreference } = await client.query(
-      `DELETE FROM language_prefs WHERE platform = $1 AND user_id = $2`,
-      [platform, userId],
-    );
-    // member_warnings (auto-moderation strikes) are keyed on raw (platform,
-    // user_id) too — a purged user's warning history goes with them.
-    const { rowCount: warnings } = await client.query(
-      `DELETE FROM member_warnings WHERE platform = $1 AND user_id = $2`,
-      [platform, userId],
-    );
-    // blocked_users (issue #572) is DELIBERATELY NOT purged here, unlike
-    // member_warnings above: forget_me/purge_user_data must never be a way to
-    // route around an admin's block by erasing the row that enforces it,
-    // including via a linked identity (SECURITY).
-    // answer_feedback (issue #118) rows this identity submitted AS RATER go
-    // with them, same as suggestions/reports above. A row where this identity
-    // was only the RECIPIENT of the rated answer is not deleted here — its
-    // interaction_id is nulled automatically by the interactions delete above
-    // via the table's ON DELETE SET NULL foreign key, leaving the rater's own
-    // helpful/unhelpful signal intact.
-    const { rowCount: answerFeedback } = await client.query(
-      `DELETE FROM answer_feedback WHERE platform = $1 AND user_id = $2`,
-      [platform, userId],
-    );
-    // knowledge_gaps (issue #208) is keyed the same way — purge coherence for
-    // anyone whose below-floor searches were logged.
-    const { rowCount: knowledgeGaps } = await client.query(
-      `DELETE FROM knowledge_gaps WHERE platform = $1 AND user_id = $2`,
-      [platform, userId],
-    );
-    // dev_team_watches (super-admin dev-team dispatches) is keyed on the same
-    // (platform, user id) identity — purge coherence for a requester's
-    // job-watch rows (which record the repo/mode/job id they dispatched).
-    const { rowCount: devTeamWatches } = await client.query(
-      `DELETE FROM dev_team_watches WHERE requester_platform = $1 AND requester_user_id = $2`,
-      [platform, userId],
-    );
-    // moderation_appeals (issue #554) is keyed the same way — purge coherence
-    // for a member's own filed appeal(s), same treatment as member_warnings.
-    const { rowCount: moderationAppeals } = await client.query(
-      `DELETE FROM moderation_appeals WHERE platform = $1 AND user_id = $2`,
-      [platform, userId],
-    );
-    // member_projects (issue #646) is keyed the same way — purge coherence for
-    // anyone who shared a project via share_project.
-    const { rowCount: memberProjects } = await client.query(
-      `DELETE FROM member_projects WHERE platform = $1 AND user_id = $2`,
-      [platform, userId],
-    );
-    // member_interests (issue #634) is keyed the same way — purge coherence
-    // for anyone who published interests via set_my_interests.
-    const { rowCount: memberInterests } = await client.query(
-      `DELETE FROM member_interests WHERE platform = $1 AND user_id = $2`,
-      [platform, userId],
-    );
-    // Member-sourced knowledge_candidates rows (issue #633, suggest_knowledge)
-    // — matched on source_platform/source_user_id, in EVERY status (pending
-    // AND accepted/declined), unlike the digest-invalidation delete above
-    // (which only removes a still-pending MACHINE row and leaves an accepted
-    // one's accountability trail intact). A member's own attributed
-    // submission is their data to erase regardless of review status; rows
-    // with source_user_id IS NULL (machine-drafted) never match this
-    // predicate, so they're untouched.
-    const { rowCount: knowledgeTips } = await client.query(
-      `DELETE FROM knowledge_candidates WHERE source_platform = $1 AND source_user_id = $2`,
-      [platform, userId],
-    );
-    // helper_notifications (issue #729, find_helper) is keyed on this
-    // identity in EITHER role — as the helper who was notified, or as the
-    // requester who triggered the notification — so both halves are deleted
-    // here, unlike every other table above which is keyed one way.
-    const { rowCount: helperNotifications } = await client.query(
-      `DELETE FROM helper_notifications
-        WHERE (helper_platform = $1 AND helper_user_id = $2)
-           OR (requester_platform = $1 AND requester_user_id = $2)`,
-      [platform, userId],
-    );
-
-    await client.query('COMMIT');
-    return (
-      (messages ?? 0) +
-      (knowledge ?? 0) +
-      (reports ?? 0) +
-      (roster ?? 0) +
-      (notes ?? 0) +
-      (suggestions ?? 0) +
-      (digestSends ?? 0) +
-      (responseStyle ?? 0) +
-      (languagePreference ?? 0) +
-      (warnings ?? 0) +
-      candidates +
-      (answerFeedback ?? 0) +
-      (knowledgeGaps ?? 0) +
-      (devTeamWatches ?? 0) +
-      (moderationAppeals ?? 0) +
-      (memberProjects ?? 0) +
-      (memberInterests ?? 0) +
-      (knowledgeTips ?? 0) +
-      (helperNotifications ?? 0)
-    );
-  } catch (err) {
-    await client.query('ROLLBACK').catch(() => {});
-    throw err;
-  } finally {
-    client.release();
-  }
-}
-
-/**
- * Delete a user's stored data: their inbound messages, the bot's replies to
- * them, knowledge entries sourced from them, content reports *they
- * submitted* as reporter, their server_roster row, admin notes kept *about*
- * them (member_notes), suggestions they filed, their response-style and
- * language preferences, answer ratings *they submitted* (issue #118), any context
- * digest built over their purged interactions, any still-pending
- * knowledge_candidates drafted from an invalidated digest (issue #102), any
- * moderation appeal(s) *they filed* (issue #554), and any knowledge_candidates
- * row *they themselves suggested* via suggest_knowledge in ANY status (issue
- * #633) — across every identity linked to them via
- * `link_member` (SECURITY: this is a deliberate blast-radius expansion —
- * linking two identities means forget_me/purge from *either* now erases
- * *both*, which is why `link_member` is CONFIRM-gated, audited, and
- * super-admin-alerted; see docs/SECURITY.md). Backs both the member-facing
- * `forget_me` and the super-admin `purge_user_data`. Membership, audit rows,
- * and reports where the user is only the *target* (not the reporter) are
- * intentionally kept (accountability data — documented in SECURITY.md).
- */
-export async function purgeUserData(platform: Platform, userId: string): Promise<number> {
-  const identities = await resolveLinkedIdentities(platform, userId);
-  let total = 0;
-  for (const identity of identities) {
-    total += await purgeSingleIdentity(identity.platform, identity.userId);
-  }
-  return total;
-}
-
-export interface MyDataSummary {
-  ownMessages: number;
-  repliesToThem: number;
-  knowledgeEntries: number;
-  reportsFiled: number;
-  suggestionsFiled: number;
-  projectsShared: number;
-  interestsPublished: number;
-  responseStyle: ResponseStyle;
-}
-
-/**
- * Read-only counterpart to `purgeSingleIdentity` — counts, rather than
- * deletes, exactly the per-table rows `forget_me`/`purge_user_data` would
- * erase for this identity (issue #188, the IPP6 access-right counterpart to
- * that deletion path), aggregated across every identity linked via
- * `link_member` the same way `purgeSingleIdentity`/`resolveLinkedIdentities`
- * already aggregate for `forget_me`. Interactions are split into the
- * caller's own messages (`user_id = $2`) and the bot's replies to them
- * (`direction = 'outbound' AND meta->>'replyToUserId' = $2`) — the same two
- * halves of `purgeSingleIdentity`'s WHERE clause, reported separately rather
- * than as one confusing lump.
- *
- * Deliberately excludes `member_notes` (issue #45: members have no
- * self-access to notes about themselves, even though `forget_me` deletes
- * them), `server_roster`, `admin_digest_sends`, `member_warnings` (already
- * self-serve via `my_warnings`), and `answer_feedback` — those stay
- * purge-only. Never add a count for any of them here to "reconcile" the
- * total with `purgeSingleIdentity`; the asymmetry is intentional.
- */
-export async function getMyDataSummary(platform: Platform, userId: string): Promise<MyDataSummary> {
-  const identities = await resolveLinkedIdentities(platform, userId);
-  let ownMessages = 0;
-  let repliesToThem = 0;
-  let knowledgeEntries = 0;
-  let reportsFiled = 0;
-  let suggestionsFiled = 0;
-  let projectsShared = 0;
-  let interestsPublished = 0;
-  for (const identity of identities) {
-    const { rows: interactionRows } = await pool.query(
-      `SELECT
-         count(*) FILTER (WHERE user_id = $2) AS own_messages,
-         count(*) FILTER (WHERE direction = 'outbound' AND meta->>'replyToUserId' = $2) AS replies_to_them
-       FROM interactions WHERE platform = $1`,
-      [identity.platform, identity.userId],
-    );
-    ownMessages += Number(interactionRows[0]?.own_messages ?? 0);
-    repliesToThem += Number(interactionRows[0]?.replies_to_them ?? 0);
-
-    // knowledge has no platform column (see purgeSingleIdentity above), so
-    // this keys on source_user_id alone, same as the DELETE it reconciles with.
-    const { rows: knowledgeRows } = await pool.query(
-      `SELECT count(*) AS n FROM knowledge WHERE source_user_id = $1`,
-      [identity.userId],
-    );
-    knowledgeEntries += Number(knowledgeRows[0]?.n ?? 0);
-
-    const { rows: reportRows } = await pool.query(
-      `SELECT count(*) AS n FROM content_reports WHERE platform = $1 AND reporter_user_id = $2`,
-      [identity.platform, identity.userId],
-    );
-    reportsFiled += Number(reportRows[0]?.n ?? 0);
-
-    const { rows: suggestionRows } = await pool.query(
-      `SELECT count(*) AS n FROM suggestions WHERE platform = $1 AND user_id = $2`,
-      [identity.platform, identity.userId],
-    );
-    suggestionsFiled += Number(suggestionRows[0]?.n ?? 0);
-
-    const { rows: projectRows } = await pool.query(
-      `SELECT count(*) AS n FROM member_projects WHERE platform = $1 AND user_id = $2 AND removed_at IS NULL`,
-      [identity.platform, identity.userId],
-    );
-    projectsShared += Number(projectRows[0]?.n ?? 0);
-
-    const { rows: interestRows } = await pool.query(
-      `SELECT count(*) AS n FROM member_interests WHERE platform = $1 AND user_id = $2`,
-      [identity.platform, identity.userId],
-    );
-    interestsPublished += Number(interestRows[0]?.n ?? 0);
-  }
-
-  return {
-    ownMessages,
-    repliesToThem,
-    knowledgeEntries,
-    reportsFiled,
-    suggestionsFiled,
-    projectsShared,
-    interestsPublished,
-    // The standing style preference isn't purge-scope data — it's a single
-    // per-identity row, so this reports the caller's own invoking identity
-    // only (same scoping set_response_style itself uses), not aggregated.
-    responseStyle: await getResponseStyle(platform, userId),
-  };
-}
-
-/**
- * Age-based retention: delete raw `interactions` older than `days`. Never
- * touches `knowledge` (curated facts are meant to be durable), `sessions`
- * (governed separately by SESSION_MAX_TURNS/_AGE_HOURS), or `admin_audit`
- * (accountability trail, retained deliberately — see SECURITY.md). Returns
- * the number of rows deleted, for operator-visible logging.
- */
-export async function purgeOldInteractions(days: number): Promise<number> {
-  const { rowCount } = await pool.query(
-    `DELETE FROM interactions WHERE created_at < now() - ($1::text || ' days')::interval`,
-    [days],
-  );
-  return rowCount ?? 0;
 }
 
 // --- Super-admin views ---------------------------------------------------------
@@ -1351,177 +957,6 @@ export async function sumBackgroundJobCosts(
   );
   const byJob = rows.map((r) => ({ job: r.job as string, costUsd: Number(r.cost) }));
   return { total: byJob.reduce((sum, r) => sum + r.costUsd, 0), byJob };
-}
-
-// --- Suggestions (member-submitted bot-improvement queue, issue #46) ---------
-
-/** Per-user cap on new suggestions within a rolling 24h window (anti-spam on the admin queue). */
-export const SUGGESTION_RATE_LIMIT_PER_DAY = 3;
-export const SUGGESTION_MAX_CHARS = 1000;
-
-export type SuggestionStatus = 'new' | 'reviewed' | 'declined' | 'done';
-
-export interface Suggestion {
-  id: number;
-  platform: Platform;
-  userId: string;
-  displayName: string | null;
-  content: string;
-  status: SuggestionStatus;
-  createdAt: Date;
-  reviewedBy: string | null;
-  reviewedAt: Date | null;
-}
-
-/**
- * Record a member's suggestion, enforcing a DB-backed rolling-24h cap per
- * user (COUNT(*) inside the insert, same restart-proof pattern as
- * createContentReport — never an in-memory or model-supplied counter).
- * Returns null when the caller is at/over the cap; the tool layer turns
- * that into a polite refusal.
- */
-export async function createSuggestion(input: {
-  platform: Platform;
-  userId: string;
-  displayName?: string;
-  content: string;
-}): Promise<{ id: number } | null> {
-  const { rows } = await pool.query(
-    `WITH recent AS (
-       SELECT count(*) AS n FROM suggestions
-        WHERE platform = $1 AND user_id = $2
-          AND created_at > now() - interval '24 hours'
-     )
-     INSERT INTO suggestions (platform, user_id, display_name, content)
-     SELECT $1, $2, $3, $4
-      WHERE (SELECT n FROM recent) < $5
-     RETURNING id`,
-    [
-      input.platform,
-      input.userId,
-      input.displayName ?? null,
-      input.content.slice(0, SUGGESTION_MAX_CHARS),
-      SUGGESTION_RATE_LIMIT_PER_DAY,
-    ],
-  );
-  return rows[0] ? { id: Number(rows[0].id) } : null;
-}
-
-/**
- * Admin-tier read of the shared suggestion queue, unscoped by submitter — a
- * member's own-only view is `listOwnSuggestions` below, not this function.
- */
-export async function listSuggestions(status?: SuggestionStatus, limit = 50): Promise<Suggestion[]> {
-  const clampedLimit = Math.min(Math.max(Math.trunc(limit) || 50, 1), 200);
-  const params: unknown[] = [];
-  let where = '';
-  if (status) {
-    params.push(status);
-    where = `WHERE status = $${params.length}`;
-  }
-  params.push(clampedLimit);
-  const { rows } = await pool.query(
-    `SELECT id, platform, user_id, display_name, content, status, created_at, reviewed_by, reviewed_at
-       FROM suggestions
-       ${where}
-      ORDER BY created_at DESC
-      LIMIT $${params.length}`,
-    params,
-  );
-  return rows.map((r) => ({
-    id: Number(r.id),
-    platform: r.platform as Platform,
-    userId: r.user_id,
-    displayName: r.display_name,
-    content: r.content,
-    status: r.status as SuggestionStatus,
-    createdAt: r.created_at,
-    reviewedBy: r.reviewed_by,
-    reviewedAt: r.reviewed_at,
-  }));
-}
-
-/**
- * Exact pending-suggestion count — a dedicated `COUNT(*)` rather than
- * `(await listSuggestions('new')).length`, which would silently understate a
- * backlog past that function's `limit` (default 50) cap, same reasoning as
- * `countAccessRequests`/`countOpenReports` (issue #133).
- */
-export async function countPendingSuggestions(): Promise<number> {
-  const { rows } = await pool.query(`SELECT count(*) AS n FROM suggestions WHERE status = 'new'`);
-  return Number(rows[0].n);
-}
-
-/**
- * Whole-day age of the oldest still-pending suggestion — the same
- * `MIN(created_at)` oldest-age mechanic `oldestAccessRequestAgeDays` (#515)
- * applies to access requests, over exactly the `status = 'new'` row set
- * `countPendingSuggestions` counts (issue #450). Guild-wide, unscoped, matching
- * its sibling count. `MIN` over an empty (all-reviewed) set is `null`, never
- * `0`, and is returned as-is so a digest reader can never mistake "no pending
- * suggestions" for "one that just arrived".
- */
-export async function oldestPendingSuggestionAgeDays(): Promise<number | null> {
-  const { rows } = await pool.query(
-    `SELECT EXTRACT(DAY FROM now() - MIN(created_at))::int AS age_days FROM suggestions WHERE status = 'new'`,
-  );
-  const ageDays = rows[0]?.age_days;
-  return ageDays === null || ageDays === undefined ? null : Number(ageDays);
-}
-
-/**
- * Self-scoped read of a member's OWN suggestions — the only member-reachable
- * read of this table (the shared queue itself stays admin-only; see the doc
- * comment on listSuggestions above). Same query shape as listSuggestions with
- * `user_id = $2` appended, the same one-predicate-append technique
- * withdrawOwnReports uses to narrow listReports's admin-scoped query down to
- * the caller's own identity.
- */
-export async function listOwnSuggestions(
-  platform: Platform,
-  userId: string,
-  limit = 10,
-): Promise<Suggestion[]> {
-  const clampedLimit = Math.min(Math.max(Math.trunc(limit) || 10, 1), 50);
-  const { rows } = await pool.query(
-    `SELECT id, platform, user_id, display_name, content, status, created_at, reviewed_by, reviewed_at
-       FROM suggestions
-      WHERE platform = $1 AND user_id = $2
-      ORDER BY created_at DESC
-      LIMIT $3`,
-    [platform, userId, clampedLimit],
-  );
-  return rows.map((r) => ({
-    id: Number(r.id),
-    platform: r.platform as Platform,
-    userId: r.user_id,
-    displayName: r.display_name,
-    content: r.content,
-    status: r.status as SuggestionStatus,
-    createdAt: r.created_at,
-    reviewedBy: r.reviewed_by,
-    reviewedAt: r.reviewed_at,
-  }));
-}
-
-/**
- * Flip a suggestion's status once triaged. Returns the resolved row's
- * platform/userId/content (so the caller can notify the submitter) or null
- * if no row matched — same "no match" signal the old boolean return gave.
- */
-export async function resolveSuggestion(
-  id: number,
-  status: Exclude<SuggestionStatus, 'new'>,
-  reviewedBy: string,
-): Promise<{ platform: Platform; userId: string; content: string } | null> {
-  const { rows } = await pool.query(
-    `UPDATE suggestions SET status = $2, reviewed_by = $3, reviewed_at = now() WHERE id = $1
-     RETURNING platform, user_id, content`,
-    [id, status, reviewedBy],
-  );
-  return rows[0]
-    ? { platform: rows[0].platform as Platform, userId: rows[0].user_id, content: rows[0].content }
-    : null;
 }
 
 // --- Question digest ---------------------------------------------------------
